+++ test/src/androidTest/kotlin/io/realm/shared/RealmObjectTests.kt
@@ -49,8 +49,8 @@ class RealmObjectTests : RealmLifeCycleTests {
     @BeforeTest
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
-        val configuration = RealmConfiguration(path = "$tmpDir/${createRandomString(16)}.realm", schema = setOf(Parent::class, Child::class))
-        realm = Realm(configuration)
+        val configuration = RealmConfiguration.defaultConfig(path = "$tmpDir/${createRandomString(16)}.realm", schema = setOf(Parent::class, Child::class))
+        realm = Realm.open(configuration)
         parent = realm.writeBlocking { copyToRealm(Parent()) }
     }